CMake: oops, compile and link `c_api.cc` when building shared library
Also use `WINDOWS_EXPORT_ALL_SYMBOLS` property to force export of
all symbols so the library is actually usable. Thanks Gera and Javi
for pointing this out.

Use of `WINDOWS_EXPORT_ALL_SYMBOL` is certainly un-Windowsish and
it's temporary. The problem is that C API does not have Windows'
`__declspec(export)` so nothing is exported buy default.

Proper solution would be either modify `c_api.cc` or provide a
`.def` file. I incline to do the former, but this only make
sense if there's any chance of this being integrated back to
upstream. Need to think of about it - meanwhile, let's live with
exporting it all.